create ClipNote component
